refactor(tabs): updated component structure
