+++ packages/react/src/components/tabs/index.ts
@@ -1,0 +1,12 @@
+export { Tab, TabList, TabPanel, TabPanels, TabsRoot } from "./tabs"
+export type {
+  TabListProps,
+  TabPanelProps,
+  TabPanelsProps,
+  TabProps,
+  TabsRootProps,
+} from "./tabs"
+
+export * from "./tabs.style"
+
+export * as Tabs from "./namespace"
