+++ packages/react/src/components/tabs/namespace.ts
@@ -1,0 +1,18 @@
+export {
+  TabsContext as Context,
+  TabsPropsContext as PropsContext,
+  TabsRoot as Root,
+  Tab,
+  TabList,
+  TabPanel,
+  TabPanels,
+  useTabsContext as useContext,
+  useTabsPropsContext as usePropsContext,
+} from "./tabs"
+export type {
+  TabsRootProps as RootProps,
+  TabListProps,
+  TabPanelProps,
+  TabPanelsProps,
+  TabProps,
+} from "./tabs"
