+++ packages/react/src/components/tabs/tabs.test.tsx
@@ -1,45 +1,45 @@
-import { Tab, TabList, TabPanel, TabPanels, Tabs } from "."
 import { a11y, render, screen } from "../../../test"
+import { Tabs } from "./"
 
 describe("<Tabs />", () => {
   test("Tabs renders correctly", async () => {
     const { container } = render(
-      <Tabs>
-        <Tab>Home</Tab>
-        <Tab>About</Tab>
-        <Tab>Contact</Tab>
+      <Tabs.Root>
+        <Tabs.Tab>Home</Tabs.Tab>
+        <Tabs.Tab>About</Tabs.Tab>
+        <Tabs.Tab>Contact</Tabs.Tab>
 
-        <TabPanel>
+        <Tabs.TabPanel>
           <p>This is home tab</p>
-        </TabPanel>
-        <TabPanel>
+        </Tabs.TabPanel>
+        <Tabs.TabPanel>
           <p>This is about tab</p>
-        </TabPanel>
-        <TabPanel>
+        </Tabs.TabPanel>
+        <Tabs.TabPanel>
           <p>This is contact tab</p>
-        </TabPanel>
-      </Tabs>,
+        </Tabs.TabPanel>
+      </Tabs.Root>,
     )
     await a11y(container)
   })
 
   test("should render tabs", async () => {
     render(
-      <Tabs>
-        <Tab>Home</Tab>
-        <Tab>About</Tab>
-        <Tab>Contact</Tab>
+      <Tabs.Root>
+        <Tabs.Tab>Home</Tabs.Tab>
+        <Tabs.Tab>About</Tabs.Tab>
+        <Tabs.Tab>Contact</Tabs.Tab>
 
-        <TabPanel>
+        <Tabs.TabPanel>
           <p>This is home tab</p>
-        </TabPanel>
-        <TabPanel>
+        </Tabs.TabPanel>
+        <Tabs.TabPanel>
           <p>This is about tab</p>
-        </TabPanel>
-        <TabPanel>
+        </Tabs.TabPanel>
+        <Tabs.TabPanel>
           <p>This is contact tab</p>
-        </TabPanel>
-      </Tabs>,
+        </Tabs.TabPanel>
+      </Tabs.Root>,
     )
 
     const tab = await screen.findByRole("tablist")
@@ -54,21 +54,21 @@ describe("<Tabs />", () => {
 
   test("should render default tab", async () => {
     render(
-      <Tabs defaultIndex={1}>
-        <Tab>Home</Tab>
-        <Tab>About</Tab>
-        <Tab>Contact</Tab>
+      <Tabs.Root defaultIndex={1}>
+        <Tabs.Tab>Home</Tabs.Tab>
+        <Tabs.Tab>About</Tabs.Tab>
+        <Tabs.Tab>Contact</Tabs.Tab>
 
-        <TabPanel>
+        <Tabs.TabPanel>
           <p>This is home tab</p>
-        </TabPanel>
-        <TabPanel>
+        </Tabs.TabPanel>
+        <Tabs.TabPanel>
           <p>This is about tab</p>
-        </TabPanel>
-        <TabPanel>
+        </Tabs.TabPanel>
+        <Tabs.TabPanel>
           <p>This is contact tab</p>
-        </TabPanel>
-      </Tabs>,
+        </Tabs.TabPanel>
+      </Tabs.Root>,
     )
 
     const aboutTabPanel = await screen.findByRole("tabpanel")
@@ -79,21 +79,21 @@ describe("<Tabs />", () => {
 
   test("should disable tab", async () => {
     render(
-      <Tabs>
-        <Tab>Home</Tab>
-        <Tab disabled>About</Tab>
-        <Tab>Contact</Tab>
+      <Tabs.Root>
+        <Tabs.Tab>Home</Tabs.Tab>
+        <Tabs.Tab disabled>About</Tabs.Tab>
+        <Tabs.Tab>Contact</Tabs.Tab>
 
-        <TabPanel>
+        <Tabs.TabPanel>
           <p>This is home tab</p>
-        </TabPanel>
-        <TabPanel>
+        </Tabs.TabPanel>
+        <Tabs.TabPanel>
           <p>This is about tab</p>
-        </TabPanel>
-        <TabPanel>
+        </Tabs.TabPanel>
+        <Tabs.TabPanel>
           <p>This is contact tab</p>
-        </TabPanel>
-      </Tabs>,
+        </Tabs.TabPanel>
+      </Tabs.Root>,
     )
 
     const aboutTab = await screen.findByRole("tab", { name: /About/i })
@@ -102,23 +102,23 @@ describe("<Tabs />", () => {
 
   test("should render custom tablist", async () => {
     render(
-      <Tabs>
-        <TabList>
-          <Tab>Home</Tab>
-          <Tab>About</Tab>
-          <Tab>Contact</Tab>
-        </TabList>
+      <Tabs.Root>
+        <Tabs.TabList>
+          <Tabs.Tab>Home</Tabs.Tab>
+          <Tabs.Tab>About</Tabs.Tab>
+          <Tabs.Tab>Contact</Tabs.Tab>
+        </Tabs.TabList>
 
-        <TabPanel>
+        <Tabs.TabPanel>
           <p>This is home tab</p>
-        </TabPanel>
-        <TabPanel>
+        </Tabs.TabPanel>
+        <Tabs.TabPanel>
           <p>This is about tab</p>
-        </TabPanel>
-        <TabPanel>
+        </Tabs.TabPanel>
+        <Tabs.TabPanel>
           <p>This is contact tab</p>
-        </TabPanel>
-      </Tabs>,
+        </Tabs.TabPanel>
+      </Tabs.Root>,
     )
 
     const tabList = await screen.findByRole("tablist")
@@ -127,23 +127,23 @@ describe("<Tabs />", () => {
 
   test("should render custom tablist (with TabPanels)", async () => {
     render(
-      <Tabs>
-        <Tab>Home</Tab>
-        <Tab>About</Tab>
-        <Tab>Contact</Tab>
+      <Tabs.Root>
+        <Tabs.Tab>Home</Tabs.Tab>
+        <Tabs.Tab>About</Tabs.Tab>
+        <Tabs.Tab>Contact</Tabs.Tab>
 
-        <TabPanels>
-          <TabPanel>
+        <Tabs.TabPanels>
+          <Tabs.TabPanel>
             <p>This is home tab</p>
-          </TabPanel>
-          <TabPanel>
+          </Tabs.TabPanel>
+          <Tabs.TabPanel>
             <p>This is about tab</p>
-          </TabPanel>
-          <TabPanel>
+          </Tabs.TabPanel>
+          <Tabs.TabPanel>
             <p>This is contact tab</p>
-          </TabPanel>
-        </TabPanels>
-      </Tabs>,
+          </Tabs.TabPanel>
+        </Tabs.TabPanels>
+      </Tabs.Root>,
     )
 
     const tabList = await screen.findByRole("tablist")
@@ -152,13 +152,13 @@ describe("<Tabs />", () => {
 
   test("Move to the previous tab with the left arrow key", async () => {
     const { user } = render(
-      <Tabs>
-        <TabList>
-          <Tab>Tab 1</Tab>
-          <Tab>Tab 2</Tab>
-          <Tab>Tab 3</Tab>
-        </TabList>
-      </Tabs>,
+      <Tabs.Root>
+        <Tabs.TabList>
+          <Tabs.Tab>Tab 1</Tabs.Tab>
+          <Tabs.Tab>Tab 2</Tabs.Tab>
+          <Tabs.Tab>Tab 3</Tabs.Tab>
+        </Tabs.TabList>
+      </Tabs.Root>,
     )
 
     const tab1 = await screen.findByRole("tab", { name: /Tab 1/i })
@@ -173,13 +173,13 @@ describe("<Tabs />", () => {
 
   test("Move to the next tab with the right arrow key", async () => {
     const { user } = render(
-      <Tabs>
-        <TabList>
-          <Tab>Tab 1</Tab>
-          <Tab>Tab 2</Tab>
-          <Tab>Tab 3</Tab>
-        </TabList>
-      </Tabs>,
+      <Tabs.Root>
+        <Tabs.TabList>
+          <Tabs.Tab>Tab 1</Tabs.Tab>
+          <Tabs.Tab>Tab 2</Tabs.Tab>
+          <Tabs.Tab>Tab 3</Tabs.Tab>
+        </Tabs.TabList>
+      </Tabs.Root>,
     )
 
     const tab1 = await screen.findByRole("tab", { name: /Tab 1/i })
@@ -192,13 +192,13 @@ describe("<Tabs />", () => {
 
   test("Move to the next tab with the down arrow key (vertical orientation)", async () => {
     const { user } = render(
-      <Tabs orientation="vertical">
-        <TabList>
-          <Tab>Tab 1</Tab>
-          <Tab>Tab 2</Tab>
-          <Tab>Tab 3</Tab>
-        </TabList>
-      </Tabs>,
+      <Tabs.Root orientation="vertical">
+        <Tabs.TabList>
+          <Tabs.Tab>Tab 1</Tabs.Tab>
+          <Tabs.Tab>Tab 2</Tabs.Tab>
+          <Tabs.Tab>Tab 3</Tabs.Tab>
+        </Tabs.TabList>
+      </Tabs.Root>,
     )
 
     const tab1 = await screen.findByRole("tab", { name: /Tab 1/i })
@@ -211,13 +211,13 @@ describe("<Tabs />", () => {
 
   test("Move to the previous tab with the up arrow key (vertical orientation)", async () => {
     const { user } = render(
-      <Tabs orientation="vertical">
-        <TabList>
-          <Tab>Tab 1</Tab>
-          <Tab>Tab 2</Tab>
-          <Tab>Tab 3</Tab>
-        </TabList>
-      </Tabs>,
+      <Tabs.Root orientation="vertical">
+        <Tabs.TabList>
+          <Tabs.Tab>Tab 1</Tabs.Tab>
+          <Tabs.Tab>Tab 2</Tabs.Tab>
+          <Tabs.Tab>Tab 3</Tabs.Tab>
+        </Tabs.TabList>
+      </Tabs.Root>,
     )
 
     const tab1 = await screen.findByRole("tab", { name: /Tab 1/i })
@@ -230,13 +230,13 @@ describe("<Tabs />", () => {
 
   test("Move to the first tab with the Home key", async () => {
     const { user } = render(
-      <Tabs>
-        <TabList>
-          <Tab>Tab 1</Tab>
-          <Tab>Tab 2</Tab>
-          <Tab>Tab 3</Tab>
-        </TabList>
-      </Tabs>,
+      <Tabs.Root>
+        <Tabs.TabList>
+          <Tabs.Tab>Tab 1</Tabs.Tab>
+          <Tabs.Tab>Tab 2</Tabs.Tab>
+          <Tabs.Tab>Tab 3</Tabs.Tab>
+        </Tabs.TabList>
+      </Tabs.Root>,
     )
 
     const tab1 = await screen.findByRole("tab", { name: /Tab 1/i })
@@ -251,13 +251,13 @@ describe("<Tabs />", () => {
 
   test("Move to the last tab with the End key", async () => {
     const { user } = render(
-      <Tabs>
-        <TabList>
-          <Tab>Tab 1</Tab>
-          <Tab>Tab 2</Tab>
-          <Tab>Tab 3</Tab>
-        </TabList>
-      </Tabs>,
+      <Tabs.Root>
+        <Tabs.TabList>
+          <Tabs.Tab>Tab 1</Tabs.Tab>
+          <Tabs.Tab>Tab 2</Tabs.Tab>
+          <Tabs.Tab>Tab 3</Tabs.Tab>
+        </Tabs.TabList>
+      </Tabs.Root>,
     )
 
     const tab1 = await screen.findByRole("tab", { name: /Tab 1/i })
