+++ packages/react/src/index.ts
@@ -93,7 +93,7 @@ export * from "./components/status"
 // export * from "./components/stepper"
 // export * from "./components/switch"
 // export * from "./components/table"
-// export * from "./components/tabs"
+export * from "./components/tabs"
 export * from "./components/tag"
 export * from "./components/text"
 export * from "./components/textarea"
