--- packages/react/src/legacy-components/tabs/index.ts
@@ -1,10 +1,0 @@
-export { Tab } from "./tab"
-export type { TabProps } from "./tab"
-export { TabList } from "./tab-list"
-export type { TabListProps } from "./tab-list"
-export { TabPanel } from "./tab-panel"
-export type { TabPanelProps } from "./tab-panel"
-export { TabPanels } from "./tab-panels"
-export type { TabPanelsProps } from "./tab-panels"
-export { Tabs } from "./tabs"
-export type { TabsProps } from "./tabs"
