--- packages/react/src/legacy-components/tabs/tabs.stories.tsx
@@ -1,593 +1,0 @@
-import type { Meta, StoryFn } from "@storybook/react"
-import { useState } from "react"
-import { Tab, TabList, TabPanel, TabPanels, Tabs } from "."
-import { Text } from "../../components/text"
-
-type Story = StoryFn<typeof Tabs>
-
-const meta: Meta<typeof Tabs> = {
-  component: Tabs,
-  title: "Components / Tabs",
-}
-
-export default meta
-
-export const Basic: Story = () => {
-  return (
-    <Tabs>
-      <Tab>孫悟空</Tab>
-      <Tab>ベジータ</Tab>
-      <Tab>フリーザ</Tab>
-
-      <TabPanel>
-        <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>へっ！きたねぇ花火だ</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>
-          私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-        </Text>
-      </TabPanel>
-    </Tabs>
-  )
-}
-
-export const WithDefaultIndex: Story = () => {
-  return (
-    <Tabs defaultIndex={1}>
-      <Tab>孫悟空</Tab>
-      <Tab>ベジータ</Tab>
-      <Tab>フリーザ</Tab>
-
-      <TabPanel>
-        <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>へっ！きたねぇ花火だ</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>
-          私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-        </Text>
-      </TabPanel>
-    </Tabs>
-  )
-}
-
-export const WithSize: Story = () => {
-  return (
-    <>
-      <Tabs size="sm">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-
-      <Tabs size="md">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-
-      <Tabs size="lg">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-    </>
-  )
-}
-
-export const WithVariant: Story = () => {
-  return (
-    <>
-      <Tabs variant="line">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-
-      <Tabs variant="sticky">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-
-      <Tabs variant="sticky-subtle">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-
-      <Tabs variant="sticky-solid">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-
-      <Tabs variant="rounded">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-
-      <Tabs variant="rounded-subtle">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-
-      <Tabs variant="rounded-solid">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-
-      <Tabs variant="unstyled">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-    </>
-  )
-}
-
-export const WithColorScheme: Story = () => {
-  return (
-    <Tabs colorScheme="red" variant="rounded-solid">
-      <Tab>孫悟空</Tab>
-      <Tab>ベジータ</Tab>
-      <Tab>フリーザ</Tab>
-
-      <TabPanel>
-        <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>へっ！きたねぇ花火だ</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>
-          私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-        </Text>
-      </TabPanel>
-    </Tabs>
-  )
-}
-
-export const WithAlign: Story = () => {
-  return (
-    <>
-      <Tabs align="start">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-
-      <Tabs align="center">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-
-      <Tabs align="end">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-    </>
-  )
-}
-
-export const WithOrientation: Story = () => {
-  return (
-    <>
-      <Tabs orientation="horizontal">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-
-      <Tabs orientation="vertical">
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-    </>
-  )
-}
-
-export const WithFitted: Story = () => {
-  return (
-    <>
-      <Tabs variant="line" fitted>
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-
-      <Tabs variant="sticky" fitted>
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </Tabs>
-    </>
-  )
-}
-
-export const Disabled: Story = () => {
-  return (
-    <Tabs>
-      <Tab>孫悟空</Tab>
-      <Tab disabled>ベジータ</Tab>
-      <Tab>フリーザ</Tab>
-
-      <TabPanel>
-        <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>へっ！きたねぇ花火だ</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>
-          私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-        </Text>
-      </TabPanel>
-    </Tabs>
-  )
-}
-
-export const Focusable: Story = () => {
-  return (
-    <Tabs>
-      <Tab>孫悟空</Tab>
-      <Tab focusable>ベジータ</Tab>
-      <Tab>フリーザ</Tab>
-
-      <TabPanel>
-        <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>へっ！きたねぇ花火だ</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>
-          私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-        </Text>
-      </TabPanel>
-    </Tabs>
-  )
-}
-
-export const Manual: Story = () => {
-  return (
-    <Tabs manual>
-      <Tab>孫悟空</Tab>
-      <Tab>ベジータ</Tab>
-      <Tab>フリーザ</Tab>
-
-      <TabPanel>
-        <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>へっ！きたねぇ花火だ</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>
-          私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-        </Text>
-      </TabPanel>
-    </Tabs>
-  )
-}
-
-export const Lazy: Story = () => {
-  return (
-    <Tabs lazy lazyBehavior="unmount">
-      <Tab>孫悟空</Tab>
-      <Tab>ベジータ</Tab>
-      <Tab>フリーザ</Tab>
-
-      <TabPanel>
-        <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>へっ！きたねぇ花火だ</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>
-          私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-        </Text>
-      </TabPanel>
-    </Tabs>
-  )
-}
-
-export const CustomTabList: Story = () => {
-  return (
-    <Tabs>
-      <TabList borderWidth={1}>
-        <Tab>孫悟空</Tab>
-        <Tab>ベジータ</Tab>
-        <Tab>フリーザ</Tab>
-      </TabList>
-
-      <TabPanel>
-        <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>へっ！きたねぇ花火だ</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>
-          私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-        </Text>
-      </TabPanel>
-    </Tabs>
-  )
-}
-
-export const CustomTabPanels: Story = () => {
-  return (
-    <Tabs variant="sticky">
-      <Tab>孫悟空</Tab>
-      <Tab>ベジータ</Tab>
-      <Tab>フリーザ</Tab>
-
-      <TabPanels borderBottomWidth={1} borderLeftWidth={1} borderRightWidth={1}>
-        <TabPanel>
-          <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>へっ！きたねぇ花火だ</Text>
-        </TabPanel>
-        <TabPanel>
-          <Text>
-            私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-          </Text>
-        </TabPanel>
-      </TabPanels>
-    </Tabs>
-  )
-}
-
-export const CustomControl: Story = () => {
-  const [index, onChange] = useState<number>(0)
-
-  return (
-    <Tabs index={index} onChange={onChange}>
-      <Tab>孫悟空</Tab>
-      <Tab>ベジータ</Tab>
-      <Tab>フリーザ</Tab>
-
-      <TabPanel>
-        <Text>クリリンのことか……クリリンのことかーーーっ！！！！！</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>へっ！きたねぇ花火だ</Text>
-      </TabPanel>
-      <TabPanel>
-        <Text>
-          私の戦闘力は530000です。ですがもちろんフルパワーであなたと戦う気はありませんからご心配なく……
-        </Text>
-      </TabPanel>
-    </Tabs>
-  )
-}
